requirement.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
faiss-cpu
pypdf